+++ Kernel/Storage/ATA/AHCI/Controller.h
@@ -10,7 +10,7 @@
 #include <AK/RefPtr.h>
 #include <AK/Types.h>
 #include <Kernel/Sections.h>
-#include <Kernel/Storage/ATA/AHCI.h>
+#include <Kernel/Storage/ATA/AHCI/Definitions.h>
 #include <Kernel/Storage/ATA/ATAController.h>
 #include <Kernel/Storage/StorageDevice.h>
