+++ Kernel/Storage/ATA/AHCI/Port.cpp
@@ -12,7 +12,7 @@
 #include <Kernel/Memory/MemoryManager.h>
 #include <Kernel/Memory/ScatterGatherList.h>
 #include <Kernel/Memory/TypedMapping.h>
-#include <Kernel/Storage/ATA/AHCIPort.h>
+#include <Kernel/Storage/ATA/AHCI/Port.h>
 #include <Kernel/Storage/ATA/ATA.h>
 #include <Kernel/Storage/ATA/ATADiskDevice.h>
 #include <Kernel/Storage/StorageManagement.h>
