+++ Kernel/Storage/ATA/ATADevice.cpp
@@ -7,8 +7,6 @@
 #include <AK/StringView.h>
 #include <Kernel/Sections.h>
 #include <Kernel/Storage/ATA/ATADevice.h>
-#include <Kernel/Storage/ATA/IDEChannel.h>
-#include <Kernel/Storage/ATA/IDEController.h>
 #include <Kernel/Storage/StorageManagement.h>
 
 namespace Kernel {
