+++ Kernel/Storage/ATA/GenericIDE/Controller.h
@@ -10,7 +10,7 @@
 #include <AK/RefPtr.h>
 #include <AK/Types.h>
 #include <Kernel/Storage/ATA/ATAController.h>
-#include <Kernel/Storage/ATA/IDEChannel.h>
+#include <Kernel/Storage/ATA/GenericIDE/Channel.h>
 #include <Kernel/Storage/StorageDevice.h>
 
 namespace Kernel {
